benchmark now uses uint64_t

+++ src/Cuckoo.cpp
@@ -25,7 +25,7 @@ void Cuckoo::Insert(std::string element) {
 	table_.Insert(element);
 }
 
-void Cuckoo::Insert(int element) {
+void Cuckoo::Insert(uint64_t element) {
 	table_.Insert(element);
 }
 
@@ -33,7 +33,7 @@ bool Cuckoo::Remove(std::string element) {
 	return table_.Remove(element);
 }
 
-bool Cuckoo::Remove(int element) {
+bool Cuckoo::Remove(uint64_t element) {
 	return table_.Remove(element);
 }
 
@@ -41,7 +41,7 @@ bool Cuckoo::Contains(std::string element) {
 	return table_.Contains(element);
 }
 
-bool Cuckoo::Contains(int element) {
+bool Cuckoo::Contains(uint64_t element) {
 	return table_.Contains(element);
 }
